requirements.txt
numpy>=1.23.5
pandas>=1.5.3
matplotlib>=3.7.1
Pillow>=9.4.0
scipy>=1.10.1
tqdm>=4.65.0
ipython>=8.10.0

# PyTorch (pip???? ????? ???)
# torch>=2.1.0
# torchvision>=0.16.0
# torchaudio>=2.1.0

opencv-python>=4.7.0.72

pyyaml>=6.0